gen-platform/app/requirements.txt
streamlit==1.38.0
requests==2.28.2
pyyaml